Move actions and constants to reducers
