+++ src/configureStore.js
@@ -1,7 +1,7 @@
 import { createStore, applyMiddleware, combineReducers } from 'redux';
 import logger from 'redux-logger';
 import thunk from 'redux-thunk';
-import {booksList} from './reducers';
+import {booksList} from './reducers/booksList';
 
 const middlewares = [thunk, logger];
